+++ src/frontend/app/services/app.service.ts
@@ -5,7 +5,8 @@ import {TourDocDataStore} from '../../shared/tdoc-commons/services/tdoc-data.sto
 import {environment} from '../../environments/environment';
 import {AppState, GenericAppService} from '@dps/mycms-commons/dist/commons/services/generic-app.service';
 import {TourDocHttpAdapter} from '../../shared/tdoc-commons/services/tdoc-http.adapter';
-import {PDocDataService} from '@dps/mycms-commons/dist/pdoc-commons/services/pdoc-data.service';
+import {StaticPagesDataService} from '@dps/mycms-commons/dist/pdoc-commons/services/staticpages-data.service';
+import {StaticPagesDataStore} from '@dps/mycms-commons/dist/pdoc-commons/services/staticpages-data.store';
 import {BaseEntityRecord} from '@dps/mycms-commons/dist/search-commons/model/records/base-entity-record';
 import {MinimalHttpBackendClient} from '@dps/mycms-commons/dist/commons/services/minimal-http-backend-client';
 import {CommonRoutingService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/common-routing.service';
@@ -98,7 +99,8 @@ export class AppService extends GenericAppService {
     };
 
     constructor(private tdocDataService: TourDocDataService, private tdocDataStore: TourDocDataStore,
-                private pdocDataService: PDocDataService, @Inject(LOCALE_ID) private locale: string,
+                private pagesDataService: StaticPagesDataService, private pagesDataStore: StaticPagesDataStore,
+                @Inject(LOCALE_ID) private locale: string,
                 private http: HttpClient, private commonRoutingService: CommonRoutingService,
                 private backendHttpClient: MinimalHttpBackendClient, private platformService: PlatformService,
                 private fallBackHttpClient: FallbackHttpClient, protected toastService: ToastrService) {
@@ -209,34 +211,48 @@ export class AppService extends GenericAppService {
         const tdocAdapter = new TourDocHttpAdapter(options);
 
         this.tdocDataStore.setAdapter('http', undefined, '', {});
-        this.pdocDataService.clearLocalStore();
+        this.pagesDataStore.setAdapter('http', undefined, '', {});
+
+        this.pagesDataService.clearLocalStore();
         this.tdocDataService.clearLocalStore();
-        this.tdocDataStore.setAdapter('http', tdocAdapter, '', {});
+
+        this.pagesDataService.setWritable(false);
+        this.tdocDataService.setWritable(false);
 
         return new Promise<boolean>((resolve, reject) => {
-            me.backendHttpClient.makeHttpRequest({ method: 'get', url: options.basePath + 'pdoc/', withCredentials: true })
+            me.backendHttpClient.makeHttpRequest({ method: 'get', url: options.basePath + 'pages/', withCredentials: true })
                 .then(function onDocsLoaded(res: any) {
                     const docs: any[] = (res['data'] || res.json());
-                    me.pdocDataService.setWritable(true);
-                    return me.pdocDataService.addMany(docs);
-                }).then(function onDocsAdded(records: BaseEntityRecord[]) {
-                    // console.log('initially loaded pdocs from server', records);
-                    me.pdocDataService.setWritable(false);
-                    me.tdocDataService.setWritable(me.appConfig.permissions.tdocWritable);
-                    return resolve(true);
-                }).catch(function onError(reason: any) {
-                    console.error('loading appdata failed:', reason);
-                    me.pdocDataService.setWritable(false);
-                    return reject(false);
-                });
+                    me.pagesDataService.setWritable(true);
+                    return me.pagesDataService.addMany(docs);
+            }).then(function onDocsAdded(records: BaseEntityRecord[]) {
+                // console.log('initially loaded pdocs from server', records);
+                me.pagesDataService.setWritable(false);
+                me.tdocDataService.setWritable(me.appConfig.permissions.tdocWritable);
+
+                me.tdocDataStore.setAdapter('http', tdocAdapter, '', {});
+
+                return resolve(true);
+            }).catch(function onError(reason: any) {
+                console.error('loading appdata failed:', reason);
+                me.pagesDataService.setWritable(false);
+
+                return reject(false);
+            });
         });
     }
 
     initStaticData(): Promise<any> {
         const me = this;
         this.tdocDataStore.setAdapter('http', undefined, '', {});
-        this.pdocDataService.clearLocalStore();
+        this.pagesDataStore.setAdapter('http', undefined, '', {});
+
+        this.pagesDataService.clearLocalStore();
         this.tdocDataService.clearLocalStore();
+
+        this.pagesDataService.setWritable(false);
+        this.tdocDataService.setWritable(false);
+
         me.appConfig.permissions.tdocWritable = false;
         me.appConfig.permissions.tdocActionTagWritable = false;
         me.appConfig.permissions.adminWritable = false;
@@ -247,7 +263,7 @@ export class AppService extends GenericAppService {
         ItemsJsDataImporter.prepareConfiguration(itemsJsConfig);
         const importer: ItemsJsDataImporter = new ItemsJsDataImporter(itemsJsConfig);
 
-        return  me.fallBackHttpClient.loadJsonPData(me.appConfig.staticPDocsFile, 'importStaticDataPDocsJsonP', 'pdocs')
+        return me.fallBackHttpClient.loadJsonPData(me.appConfig.staticPDocsFile, 'importStaticDataPDocsJsonP', 'pdocs')
             .then(function onPDocLoaded(data: any) {
                 if (data['pdocs']) {
                     return Promise.resolve(data['pdocs']);
@@ -255,11 +271,13 @@ export class AppService extends GenericAppService {
 
                 return Promise.reject('No static pdocs found');
             }).then(function onPDocParsed(docs: any[]) {
-                me.pdocDataService.setWritable(true);
-                return me.pdocDataService.addMany(docs);
+                me.pagesDataService.setWritable(true);
+                return me.pagesDataService.addMany(docs);
             }).then(function onPDocsAdded(pdocs: BaseEntityRecord[]) {
                 console.log('initially loaded pdocs from assets', pdocs);
-                me.pdocDataService.setWritable(false);
+
+                me.pagesDataService.setWritable(false);
+                me.tdocDataService.setWritable(true);
 
                 const promises = [];
                 for (const staticTDocsFile of me.appConfig.staticTDocsFiles) {
@@ -313,7 +331,8 @@ export class AppService extends GenericAppService {
                 return Promise.resolve(true);
             }).catch(function onError(reason: any) {
                 console.error('loading appdata failed:', reason);
-                me.pdocDataService.setWritable(false);
+
+                me.pagesDataService.setWritable(false);
                 me.tdocDataService.setWritable(false);
 
                 return Promise.reject(false);